Add files via upload
It's a file for checking database are already exist or not. We have created a function with try & accept in python for checking database are already exist or not, if database are already exist then it will not create a new database and program will not fail. If database are not exist it will create a Database access_control_system.